+++ CHANGES.md
@@ -6,6 +6,8 @@
 - Remove returned bitvector in ops, user is now required to track it themselves
 - Use `~dst` instead of `~result`
 - Add `Allocate` variant of all ops in a submodule
+- Allow equality check between different-length bitvectors
+
 
 ## 0.0.5
 - 32-bit compatible